Separate order by integration tests from query builder test
