+++ QueryBuilder_Cassandra/src/test/java/net/sf/esfinge/querybuilder/cassandra/integration/main/CassandraSimpleQueryBuilderTest.java
@@ -14,7 +14,7 @@ import java.util.List;
 
 import static org.junit.Assert.*;
 
-public class CassandraQueryBuilderIntegrationTest extends CassandraBasicDatabaseTest {
+public class CassandraSimpleQueryBuilderTest extends CassandraBasicDatabaseTest {
 
     CassandraTestQuery testQuery = QueryBuilder.create(CassandraTestQuery.class);
 
@@ -67,23 +67,6 @@ public class CassandraQueryBuilderIntegrationTest extends CassandraBasicDatabase
         List<Person> list = testQuery.getPersonByNameOrLastName("Pedro", "Silva");
     }
 
-    /*@Test
-     TODO: IMPLEMENT JOINS WITH OTHER TABLES
-    public void queryWithOtherTable(){
-        List<Person> list = tq.getPersonByAddressCity("Juiz de Fora");
-        assertEquals("The list should have 2 persons", 2, list.size());
-        assertEquals("The first should be Antonio", "Antonio", list.get(0).getName());
-        assertEquals("The second should be Silvia", "Silvia", list.get(1).getName());
-    }
-
-    @Test
-    public void compositeQueryWithOtherTable(){
-        List<Person> list = tq.getPersonByLastNameAndAddressState("Silva","SP");
-        assertEquals("The list should have 2 persons", 2, list.size());
-        assertEquals("The first should be Pedro", "Pedro", list.get(0).getName());
-        assertEquals("Marcos", list.get(1).getName());
-    }*/
-
     @Test
     public void queryWithGreaterThanTest() {
         List<Person> list = testQuery.getPersonByAge(40);
@@ -123,54 +106,5 @@ public class CassandraQueryBuilderIntegrationTest extends CassandraBasicDatabase
         List<Person> list = testQuery.getPersonByIdAndName(1, "Homer", 48);
     }
 
-    @Test
-    public void orderByQueryWithOneFieldTest() {
-        List<Person> list = testQuery.getPersonOrderByName();
-
-        String[] actualNames = list.stream().map(p -> p.getName()).toArray(String[]::new);
-        String[] expectedNames = {"Antonio", "Marcos", "Maria", "Pedro", "Silvia"};
-
-        assertArrayEquals(expectedNames, actualNames);
-    }
-
-    @Test
-    public void orderByQueryWithOneFieldAndParameterDescendentTest() {
-        List<Person> list = testQuery.getPersonByAgeOrderByNameDesc(21);
-
-        String[] actualNames = list.stream().map(p -> p.getName()).toArray(String[]::new);
-        String[] expectedNames = {"Maria", "Marcos", "Antonio"};
-
-        assertArrayEquals(expectedNames, actualNames);
-    }
-
-    @Test
-    public void orderByQueryWithTwoFieldsTest() {
-        List<Person> list = testQuery.getPersonOrderByLastNameAndName();
-
-        String[] actualNames = list.stream().map(p -> p.getName()).toArray(String[]::new);
-        String[] expectedNames = {"Silvia", "Maria", "Antonio", "Marcos", "Pedro"};
-
-        assertArrayEquals(expectedNames, actualNames);
-    }
-
-    @Test
-    public void orderByQueryWithTwoFieldsWithOrderingTest() {
-        List<Person> list = testQuery.getPersonOrderByLastNameDescAndNameAsc();
-
-        String[] actualNames = list.stream().map(p -> p.getName()).toArray(String[]::new);
-        String[] expectedNames = {"Marcos", "Pedro", "Antonio", "Maria", "Silvia"};
-
-        assertArrayEquals(expectedNames, actualNames);
-    }
-
-    @Test
-    public void complexOrderByQueryTest() {
-        List<Person> list = testQuery.getPersonByAgeAndLastNameOrderByAgeAndLastNameDescAndName(51, "Silva");
-
-        String[] actualNames = list.stream().map(p -> p.getName()).toArray(String[]::new);
-        String[] expectedNames = {"Pedro", "Marcos"};
-
-        assertArrayEquals(expectedNames, actualNames);
-    }
 
 }
